+++ integration_tests/suite/wizard/test_wizard.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import copy
@@ -8,7 +8,6 @@ import re
 from hamcrest import (
     assert_that,
     contains_inanyorder,
-    equal_to,
     empty,
     has_entry,
     has_entries,
@@ -266,7 +265,43 @@ class TestWizardDefaultValue(IntegrationTest):
             assert_that(queries.sccp_has_language('en_US'))
 
 
-class TestWizard(IntegrationTest):
+class TestWizardGet(IntegrationTest):
+    @mocks.sysconfd()
+    @mocks.auth()
+    def test_get(self, sysconfd, auth):
+        data = copy.deepcopy(COMPLETE_POST_BODY)
+
+        response = self.confd.wizard.get()
+        assert_that(
+            response.item,
+            has_entries(
+                configured=False,
+                configurable=True,
+                configurable_status=has_entries(
+                    {'wazo-auth': 'ok', 'wazo-provd': 'ok'}
+                ),
+            ),
+        )
+
+        response = self.confd.wizard.post(data)
+        response.assert_ok()
+
+        response = self.confd.wizard.get()
+        assert_that(response.item, has_entries(configured=True))
+
+        self.stop_service('provd')
+
+        response = self.confd.wizard.get()
+        assert_that(
+            response.item,
+            has_entries(
+                configurable=False,
+                configurable_status=has_entries({'wazo-provd': 'fail'}),
+            ),
+        )
+
+
+class TestWizardPost(IntegrationTest):
     @mocks.sysconfd()
     @mocks.auth()
     def test_post(self, sysconfd, auth):
@@ -274,13 +309,13 @@ class TestWizard(IntegrationTest):
         bus_events = self.bus.accumulator('config.wizard.created')
 
         response = self.confd.wizard.get()
-        assert_that(response.item, equal_to({'configured': False}))
+        assert_that(response.item, has_entries(configured=False))
 
         response = self.confd.wizard.post(data)
         response.assert_ok()
 
         response = self.confd.wizard.get()
-        assert_that(response.item, equal_to({'configured': True}))
+        assert_that(response.item, has_entries(configured=True))
 
         self.validate_db(data)
         self.validate_auth(auth, data)
@@ -433,13 +468,13 @@ class TestWizardSteps(IntegrationTest):
         bus_events = self.bus.accumulator('config.wizard.created')
 
         response = self.confd.wizard.get()
-        assert_that(response.item, equal_to({'configured': False}))
+        assert_that(response.item, has_entries(configured=False))
 
         response = self.confd.wizard.post(data)
         response.assert_ok()
 
         response = self.confd.wizard.get()
-        assert_that(response.item, equal_to({'configured': True}))
+        assert_that(response.item, has_entries(configured=True))
 
         self.validate_auth(auth, data)
         self.validate_sysconfd(sysconfd, data)