+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 20.02
+
+* The GET on `1.1/wizard` now returns the `configurable` and `configurable_status`' fields
+
 ## 20.01
 
 * The voicemail `name` no longer change according to the user `firstname`/`lastname`